Add --host=0.0.0.0 if running llama.cpp serve within a container
Turn on some testing of --nocontainer serve and run, at least with
dryrun.

+++ test/system/030-run.bats
@@ -3,31 +3,37 @@
 load helpers
 
 @test "ramalama --dryrun run basic output" {
-    skip_if_nocontainer
-
     model=tiny
     image=m_$(safename)
 
-    run_ramalama info
-    conman=$(jq .Engine <<< $output | tr -d '"' )
-    verify_begin="${conman} run --rm -i --label RAMALAMA --security-opt=label=disable --name"
+    if is_container; then
+	run_ramalama info
+	conman=$(jq .Engine <<< $output | tr -d '"' )
+	verify_begin="${conman} run --rm -i --label RAMALAMA --security-opt=label=disable --name"
 
-    run_ramalama --dryrun run ${model}
-    is "$output" "${verify_begin} ramalama_.*" "dryrun correct"
-    is "$output" ".*${model}" "verify model name"
+	run_ramalama --dryrun run ${model}
+	is "$output" "${verify_begin} ramalama_.*" "dryrun correct"
+	is "$output" ".*${model}" "verify model name"
 
-    run_ramalama --dryrun run --name foobar ${model}
-    is "$output" "${verify_begin} foobar .*" "dryrun correct with --name"
-    is "$output" ".*${model}" "verify model name"
+	run_ramalama --dryrun run --name foobar ${model}
+	is "$output" "${verify_begin} foobar .*" "dryrun correct with --name"
+	is "$output" ".*${model}" "verify model name"
 
-    run_ramalama --dryrun run --name foobar ${model}
-    is "$output" "${verify_begin} foobar .*" "dryrun correct with --name"
+	run_ramalama --dryrun run --name foobar ${model}
+	is "$output" "${verify_begin} foobar .*" "dryrun correct with --name"
 
-    run_ramalama 1 --nocontainer run --name foobar tiny
-    is "${lines[0]}"  "Error: --nocontainer and --name options conflict. --name requires a container." "conflict between nocontainer and --name line"
+	run_ramalama 1 --nocontainer run --name foobar tiny
+	is "${lines[0]}"  "Error: --nocontainer and --name options conflict. --name requires a container." "conflict between nocontainer and --name line"
 
-    RAMALAMA_IMAGE=${image} run_ramalama --dryrun run ${model}
-    is "$output" ".*${image} /bin/sh -c" "verify image name"
+	RAMALAMA_IMAGE=${image} run_ramalama --dryrun run ${model}
+	is "$output" ".*${image} /bin/sh -c" "verify image name"
+    else
+	run_ramalama --dryrun run ${model}
+	is "$output" 'llama-cli -m /path/to/model --in-prefix --in-suffix --no-display-prompt -p.*' "dryrun correct"
+
+	run_ramalama 1 run --name foobar tiny
+	is "${lines[0]}"  "Error: --nocontainer and --name options conflict. --name requires a container." "conflict between nocontainer and --name line"
+    fi
 }
 
 @test "ramalama run tiny with prompt" {
